stateless.bbclass: STATELESS_SRC for injecting source code and patches
Modifying SRC_URI and the required SRC_URI[*.sha256sum] in a global
stateless.inc triggers world rebuilds because the varflags are shared
between all recipes.

STATELESS_SRC sets SRC_URI separately in each recipe and thus avoids
that. It's also slightly easier to use because the name tying the
shasum to the item is generated automatically.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>