stateless.bbclass: "factory" mode for moving /etc entries
/etc does not need to be empty at runtime. It's okay to populate
it from OS defaults during the early boot phase. systemd already
supports such a "factory" reset with tmpfiles.d.

STATELESS_MV_ROOTFS = "foo=factory" moves /etc/foo into the
factory defaults and ensures that systemd re-creates its.

TODO: this is done with a single stateless.conf file, which won't work
for the case where individual recipes want to use this feature. Either
fix this or remove support for it in the recipe case.

/etc needs to be populated before services using the content are
started. In some cases, additional systemd dependencies are needed
to ensure that, for example for ld.so.conf with:
  STATELESS_MV_ROOTFS = "ld.so.conf=factory:ldconfig.service"

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>