refkit images: mount rw
refkit.conf and refkit-image.bbclass had conflicting settings: the
rootfs was mounted ro so that fsck could be run, but that got disabled
in the distro.

Disabling the check still makes sense, but is better done in the image
class, and then we can also mount read/write immediately. That is
better for a stateless boot, too.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>