ostree-{config,image}.bbclass: add classes for OSTree image creation.
Added classes for creating an OSTree-enabled rootfs, image and the
corresponding OSTree repositories. The image class defines an ostree
image variant which can be used to generate an OSTree-enabled version
of its base image. It relies on mk-ostree.sh to do all the heavy lift-
ing. The boot-time OSTree rootfs (version) selection glue is handled
by the ostree initramfs module.

The ostree-config class defines and splits out the visible OSTree-
related variables. A subset of these can be overridden by the user.
They are split out in case we end up having packages that need to be
aware of the OSTree configuration details.

Signed-off-by: Krisztian Litkey <kli@iki.fi>