ovmf-shell-image: create a key-setter derivative
runqemu in OEQA framework cannot be used to command EFI shell
because of the built-in assumptions that the system must boot to
a getty login.

Because of this, we must be able to enroll EFI Secure Boot keys
automatically. This is done with a special ovmf-shell-image-enrollkeys
image that automatically launches startup.nsh script and deploys the keys
in ovmf.secboot.

Signed-off-by: Mikko Ylinen <mikko.ylinen@linux.intel.com>