scripts/mk-ostree.sh: add mk-ostree.sh for creating an OSTree rootfs.
mk-ostree.sh takes a pristine image rootfs, shuffles its layout around
to be OSTree-compatible, commits it into a per-image primary OSTree
bare-user repository, and creates a final OSTree-enabled rootfs by
checking out the rootfs (or 'deploying' in OSTree terminology) from
this repository.

Additionally, the script can publish the generated rootfs in another
repository in archive-z2 format which then can be exposed over HTTP
and used by clients to upgrade themselves using OSTree.

If a GPG key(ring and/or ID) is provided it will be used to sign the
commits in the repository.

Signed-off-by: Krisztian Litkey <kli@iki.fi>