.gitmodules: pull submodules from the main repo
This relies on having the relevants commits set up in the main repo so
that cloning that also fetches the submodule commits. This was done by
manually creating "refs/heads/modules/<submodule>" branches that
mirror each submodule.

The relative URL continues to work when forking the main repo. But
because "url = ./" isn't valid (leads to "...git/" urls that are
invalid) the repo name needs to be repeated and thus the file needs
to be edited when forking and publishing under a different name.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>