meta-refkit-core: remove gsignond and dependencies
Some parts of gsignond depend on recipes which are not in
OE-core. These dependencies must be made optional for gsignond to be
hosted in the upcoming meta-refkit-core. As there is no current need
for it, the set of recipes get removed instead.

A layer like meta-security might be a better place for gsignond
anyway. ecryptfs-utils was already proposed for it by Wind River
("[yocto] [meta-security][PATCH 2/2] ecryptfs-utils: add new recipe").

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>