intel-iot-refkit: adapt Docker container to build instead of the other way around
Previously we manipulated the HOSTTOOLS variable to get the build to
succeed in a minimal OpenSUSE installation. This happened to work, but
caused problems for a selftest that didn't use the refkit-ci.inc file
and might also break once the tools really become necessary.

So better adapt the container so that it provides what is expected.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>