refkit: replace development/production .inc files with per-image IMAGE_MODE
Forcing developers to choose between production and development builds
was non-standard (thus tripping up new developers) and inflexible
(couldn't build product and development images at the same time).

The new image-mode.bbclass replaces that with a per-image IMAGE_MODE
variable that can be set as desired in local.conf for existing image
recipes or fixed on a specific value in specific images. It also
customizes /etc/motd in the images, just like the refkit-specific code
did before.

The default (in our local.conf.sample) is to build images in
development mode. This is a bit more relaxed than the approach we
inherited from Ostro OS where a developer was explicitly forced to
choose.

Because setting the Secure Boot signing keys no longer depends on the
.inc files, the default development keys can be moved to
meta-refkit-core, where they are used as default for IMAGE_MODE !=
production.

It is possible that the same build may now need an refkit-initramfs in
development mode (= default dm-verity keys) and one in production
mode. This is solved by providing refkit-initramfs in different variants,
generated automatically using image-mode-variants.bbclass. For this to
work, we have to set valid refkit modes globally.

Extra care is needed to avoid setting IMAGE_MODE and IMAGE_MODE_VALID
themselves globally: the values that are valid for refkit images are
not valid for other images, like core-image-minimal. While that
particular image doesn't do anything with IMAGE_MODE, others might.
Therefore we set REFKIT_IMAGE_MODE[_VALID] globally and copy to the
values checked by the helper classes in the refkit images.

The refkit images use the REFKIT_IMAGE_MODE set in our
local.conf.sample, which may vary in actual builds. Testing continues
to use that because if already built in the right mode, there's no
need to rebuilt an image with a fixed mode.

If someone wants, they now can also built images with a specific
mode that also gets included in the resulting image files (for example,
refkit-image-computervision-development).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>