meta-refkit-core: clean up refkit-config
The refkit-conf.inc is now split into distinct sections, with more
general ones at the top and specific recipe changes at the bottom.
PACKAGECONFIG changes are meant to be done in the .inc file, so move
the systemd change there.

More complex changes are meant to be in .bbappends, so recreate the
python .bbappends (but with an .inc file for the common code this
time, and in a simpler way that can be made conditional).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>