refkit-image.bbclass: honor EFI_PROVIDER
meta-intel sets a WKS_FILE (systemd-bootdisk-uuid.wks) which seems to
assume that the class in EFI_PROVIDER (rmc-boot) gets inherited by the
images.

This used to be done indirectly through the "live" IMAGE_FSTYPES
entry, but doesn't not happen in refkit images because we disable that
type. Instead we need to load that class when our own
image-dsk.bbclass is not active.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>