refkit: remove syslinux and grub-efi only in the distro
Disabling the dependency on syslinux and grub-efi is only valid
for refkit images and (besides saving some resources) necessary
when using refkit.conf because we do not list those components
as supported.

When using other distros, we need to keep these dependencies,
because otherwise non-refkit images built without errors but
then do not boot as no bootx64.efi gets installed.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>