selftest/refkit-poky.py: build and boot common image
With several minor problems fixed, we can now build and boot the
common image both with the default Poky
configuration (test_common_poky_config) and with DISTRO=poky +
refkit-config (test_common_refkit_config).

This replaces build-testing of the computer vision image. The problems
found with that test were all in common code, so it makes sense to
keep the test a bit simpler and focus just on that.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>