refkit-supported-recipes.txt: support refkit images and packagegroups
Not listing them had the effect of excluding them from a world build.
Explicitly building them was possible without triggering warnings.

We want our package groups to be part of a world build to detect
dependency errors. The same goes for images: they are excluded
because of EXCLUDE_FROM_WORLD in image.bbclass, but a test might
want to change that.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>