meta-refkit: bump bblayers.conf version
meta-refkit was added, so alert users that they need to update
their bblayers.conf by bumping the version number.

Version numbers >= 4 and <= 7 have a special meaning in OE-core's
sanity.bbclass in the code which updates bblayers.conf. We inherit
that code and cannot disable it (currently; patch for OE-core in
preparation).

By bumping directly to 8 we ensure that the OE-core update code at
least eventually aborts and prints the desired error message (for some
reasons, twice). However, the output now is a bit confusing when
updating from bblayers.conf with version 4:

  NOTE: Your conf/bblayers.conf has been automatically updated.
  NOTE: Your conf/bblayers.conf has been automatically updated.
  NOTE: Your conf/bblayers.conf has been automatically updated.
  ERROR: Your version of bblayers.conf has the wrong LCONF_VERSION (has 7, expecting 8).
  Please compare your file against bblayers.conf.sample and merge any changes before continuing.
  "meld conf/bblayers.conf /fast/work/intel-iot-refkit//meta-refkit/conf/bblayers.conf.sample"

  is a good way to visualise the changes.
  ERROR: Your version of bblayers.conf has the wrong LCONF_VERSION (has 7, expecting 8).
  Please compare your file against bblayers.conf.sample and merge any changes before continuing.
  "meld conf/bblayers.conf /fast/work/intel-iot-refkit//meta-refkit/conf/bblayers.conf.sample"

  is a good way to visualise the changes.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>