meta-flatpak: use "flatpak" branch without ostree system update
Developing OSTree system update support for refkit in the meta-flatpak
repository and layer was a shortcut. The right solution would be
either a distro-agnostic "meta-ostree" layer or distro-specific OSTree
system update support in the meta-refkit-core layer.

The current code is very specific to refkit, and it is unclear how
generic it can be made. For now the focus is to develop a solution for
refkit (i.e. relying on systemd/UEFI/initramfs-framework, and
supporting the refkit disk layout with a UEFI combo app), so all files
from the "ostree" branch were moved into meta-refkit-core and we go
back to the "flatpak" branch of meta-flatpak.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>