refkit-image.bbclass: enable emergency shell in development mode
The idea behind imagetype=development is good, having a development
initramfs fall back into a shell on fatal errors is useful. However,
now that the have IMAGE_MODE=development, we should just configure the
kernel command line accordingly directly. That way we also get an
emergency shell in other cases, not just for OSTree failures (which is
where the mode had been checked).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>