refkit-image.bbbclass: single .wks with and without OSTree
The two .wks files had drifted apart in comments. We can avoid that by
only maintaining a single .wks in which we dynamically decided from
where the rootfs comes.

This also undos a shortcoming of the "refkit-image.bbbclass: ostree
support" where image-dsk.bbclass had to be modified to add ostree
support.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>