ostree: include MACHINE and image PN in branch name
We have to use the machine name instead of the architecture name as
part of the branch, because otherwise different machines with
different machine-specific files end up sharing the same branch when
they share the same architecture.

We can still share the same architecture-specific files by putting
more than one image into the same repository. For this to work,
the branch name has to include the image name.

This way it should also be possible to switch back and forth between
different images on the same hardware.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>