ostree: remove invalid (?) reuse of old work repo
It is not clear under which specific conditions bailing out of
do_ostree_prepare_rootfs() was valid. I ran into cases where the
ostree.rootfs was missing because of this check.

In addition, it prevents re-generating ostree.rootfs and the image
after making changes to the OSTree code.

Therefore now do_ostree_prepare_rootfs() produces its output anew when
it gets triggered.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>