ostree: re-run tasks when scripts change
We teach bitbake about the task dependency on the content of the
scripts that it calls. This way, editing the scripts and invoking
bitbake again will reuse the existing rootfs and just run the creation
of rootfs.ostree, followed by the normal wic image creation and
publishing.

We don't need to care about the location of the scripts changing,
that's already covered by the normal signature calculation of the task
method.

See also https://bugzilla.yoctoproject.org/show_bug.cgi?id=11605

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>