ostree: generate .pub/.sec GPG key pair in TOPDIR
Previously, the location was unspecified and happened to depend
on where the task was invoked (TOPDIR?). When invoking the
run.do_ostree_prepare_rootfs script under devshell, it happened to be
something else.

Now we consistently enforce a "cd $TOPDIR". This is a stop-gap
measure: the OSTREE_GPGDIR might be set differently for different
images, so putting its keys under TOPDIR might cause conflicts. It
would be better if gpg-keygen.sh and its users were changed to have
that key pair also under OSTREE_GPGDIR. This is a change that'll
affect multiple different repos and/or branches.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>