ostree: fix initialization of remote in new images
The new image must have a remote configured with the right URL and the
matching GPG key.

The GPG part was previously missing, and "ostree admin" failed because
it wasn't running under OSTree ("error: Not currently booted into an
OSTree system").

The script was also checking the wrong variable ("OSTREE_REMOTE" is
not what the parameter sets).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>