ostree: system update testing
This consists of a generic part (systemupdatebase.py) and a part that
is specific to the integration of OSTree into
refkit (refkit_ostree.py).

Making the OSTree repo available to the virtual machine guest has to
jump through some hoops because only slirp networking works. It
currently relies on netcat "so that QEMU behaves similar to an inetd
process for that virtual server" (quote from the QEMU documentation).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>