system update: test with two image recipes
With OSTree it is fairly easy to build original and update in
different image recipes. This makes it easier to debug the test,
because both build directories are preserved. It might also speed up
running the tests repeatedly, because the rootfs should be reusable.

TODO: in practice, we still need to rebuild. Check why.

The base infrastructure could be used to test switching between
different images with OSTree.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>