systemd-sysusers.bbclass: delete converted files
Once we have added users and groups to the /etc databases, the systemd
sysusers.d config files shouldn't have any effect any more at runtime
and thus can be removed from images.

This works around a limitation of the current implementation and/or a
bug in the base files: because the user "nobody" exists, the current
implementation does nothing. But there is no group "nobody", and
systemd then adds that at runtime. That prevents updating /etc/group
with OSTree, because the file is always considered as "locally
modified by admin".

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>