refkit: update meta-security
The TPM support was split out into its own layer. It now uses dynamic
layer dependencies, including meta-xfce when x11 is enabled. We
therefore must also add meta-xfce to our default layers, to ensure
that builds with x11 enabled work.

Because yocto-compat-layer.py does not handle such dynamic
dependencies (and perhaps should have to?), we have to add the
dependencies in advance.  See "[yocto] dynamic layer dependencies in
meta-security" for a discussion of the problem.

* meta-security 3bcca12...89683b4 (22):
  > apparmor: fix python packaging issue
  > ecryptfs-utils: add new recipe
  > keyutils: add new recipe
  > layer.conf: fix typo (meta-filesystems vs mete-filesystems)
  > packagegroup-security-tpm-i2c: fix syntax
  > tor: add recipe
  > layer-conf: Use *_FEATURES in LAYERDEPENDS
  > security-core package group: add few more apps
  > clamav: fix new build error
  > sssd: update SRC_URI as git.fedorahosted.org shut down
  > tpm2: package groups fixes
  > linux-stable: fix module selections
  > tpm-image: used for testing for now.
  > kernel tpm rework
  > tpm-i2c: some systems us i2c TPM
  > tpm packagegroups: split into logical units
  > packagegroup: remove tpm components
  > meta-tpm: add base package group as was in meta-security
  > change tpm from distro to machine feature
  > tpm: move to a sub layer
  > swtpm-wrappers-native.bb: need netstat
  > freediameter: Add recipe

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>